requirements.txt
django
djangorestframework
langchain
langchain-community
pypdf
sentence_transformers
faiss-cpu
chromadb